fix: next build errors in blog
