+++ content/posts/fermi-problems.mdx
@@ -3,7 +3,8 @@ title: Fermi Problems
 date: 2025-01-08
 imageUrl: /space.png
 description: Estimation problems
-tags: ['physics', 'problem-solving', 'estimation', 'science', 'mathematics']
+tags:
+    - problem-solving
 ---
 
 Sometimes it's fun to try to calculate without knowing everything
