+++ content/posts/wins-tracker.mdx
@@ -4,13 +4,7 @@ date: 2025-01-02
 imageUrl: /e7_sunset.jpeg
 description: Keeping track of the good moments
 tags:
-    [
-        'productivity',
-        'mental-health',
-        'personal-development',
-        'habits',
-        'reflection',
-    ]
+    - reflection
 ---
 
 I've begun to track my daily wins. Anything that makes me feel like I got something done, it goes in. Any sort of accomplishment, or just random things that happened to go well today.
