+++ content/posts/zero-latency.mdx
@@ -4,13 +4,7 @@ date: 2025-01-01
 imageUrl: /stars.png
 description: Minimizing thought latency
 tags:
-    [
-        'productivity',
-        'personal-development',
-        'discipline',
-        'motivation',
-        'habits',
-    ]
+    - productivity
 ---
 
 > Imagine you're on your deathbed—and standing around your bed are the ghosts representing your unfilled potential. The ghosts of the ideas you never acted on. The ghosts of the talents you didn't use. And they're standing around your bed. Angry. Disappointed. Upset. 'We came to you because you could have brought us to life,'they say. 'And now we go to the grave together.' So I ask you today: How many ghosts are going to be around your bed when your time comes?
